move error model to viewmodels
